Add files via upload
The first surge of Covid- 19 impacted the global frugality as the world was no way ready for the epidemic. It redounded in a rise in cases, a rise in deaths, a rise in severance and a rise in poverty, performing in an profitable retardation. Using python language,  the spread of Covid cases and its impact on frugality was dissected.